Add note about discouraged monthly tags
Stable tags are heavily encouraged on Docker's infrastructure to build
reproducible images. But as ArchLinux does not provide a stable release
with a stable mirror, it's not possible to only install the necessary
packages in a Docker image and rebuild it a few months later.

A full system **upgrade** is always necessary, and therefore a monthly
tag will increase the image size, but won't make any difference to the
image, if built with `latest` tag.

See #21